chore: tidy files
